+++ src/clj_annotations/validation.clj
@@ -127,9 +127,11 @@
 (defn- validate-vector-attribute
   [path schema obj {:keys [make-result validation-fns] :as opts}]
   (if (sequential? obj)
-    (letfn [(validate [i v]
-              (validate-scalar-attribute (conj path i) schema v opts))]
-     (apply concat (map-indexed validate obj)))
+    (let [vec-result (eval-validity-condition path schema obj opts)
+          validate   (fn [i v]
+                       (validate-scalar-attribute (conj path i) schema v opts))
+          elem-results (apply concat (map-indexed validate obj))]
+      (concat vec-result elem-results))
     (make-result schema obj path :non-array-value nil)))
 
 (defn- validate-attribute